refactor: change name recordFile to recordFileManipulator
